IVE BEEN BUSY

+++ include/PR/ultratypes.h
@@ -16,7 +16,8 @@ typedef signed int             s32;
 typedef unsigned int           u32;
 
 typedef float  f32;
-typedef double f64;
+///typedef double f64;
+typedef float f64;
 
 #ifdef TARGET_N64
 typedef signed long long int   s64;
