+++ src/game/geo_misc.c
@@ -47,9 +47,9 @@ void make_vertex(Vtx *vtx, s32 n, s16 x, s16 y, s16 z, s16 tx, s16 ty, u8 r, u8 
 #else
 void make_vertex(Vtx *vtx, s32 n, f32 x, f32 y, f32 z, s16 tx, s16 ty, u8 r, u8 g, u8 b, u8 a) {
 #endif
-    vtx[n].v.ob[0] = x;
-    vtx[n].v.ob[1] = y;
-    vtx[n].v.ob[2] = z;
+    vtx[n].v.ob[0] = (s16)x;
+    vtx[n].v.ob[1] = (s16)y;
+    vtx[n].v.ob[2] = (s16)z;
 
     vtx[n].v.flag = 0;
